split preprocessing into 2 files
